moved the api fetch to reusable functions
