+++ src/pages/ProductPage/index.tsx
@@ -1,48 +1,35 @@
-import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import setPrice from '../../features/price/setPrice'
 import { useShoppingCart } from '../../context/ShoppingCartContext'
 import * as s from './index.styled'
-
-const url = 'https://api.noroff.dev/api/v1/online-shop'
+import { SingleProductFetch } from '../../data/SingleProductFetch'
 
 function ProductPage() {
-  const { id } = useParams()
-  const [product, setProduct] = useState<any>([])
+  const { id } = useParams() as any
   const { increaseQuantity } = useShoppingCart()
+  const product = SingleProductFetch(id)
 
-  useEffect(() => {
-    async function getData() {
-      const response = await fetch(url + '/' + id)
-      const json = await response.json()
-      setProduct(json)
-    }
-    getData()
-  }, [])
-
-  function renderReviews() {  
-
-      return (
-        <div>
-          <h3>reviews</h3>
-          <ul>
-            {product.reviews &&
-              product.reviews.map(
-                (review: {
-                  id: string
-                  username: string
-                  rating: number
-                  description: string
-                }) => (
-                  <li key={review.id}>
-                    {review.username} {review.rating}/5 {review.description}
-                  </li>
-                )
-              )}
-          </ul>
-        </div>
-      )
-           
+  function renderReviews() {
+    return (
+      <div>
+        <h3>reviews</h3>
+        <ul>
+          {product.reviews &&
+            product.reviews.map(
+              (review: {
+                id: string
+                username: string
+                rating: number
+                description: string
+              }) => (
+                <li key={review.id}>
+                  {review.username} {review.rating}/5 {review.description}
+                </li>
+              )
+            )}
+        </ul>
+      </div>
+    )
   }
 
   return (
@@ -52,7 +39,9 @@ function ProductPage() {
         <h1>{product.title}</h1>
         <p>{product.description}</p>
         {setPrice(product)}
-        <button onClick={() => increaseQuantity(product.id, product.discountedPrice)}>
+        <button
+          onClick={() => increaseQuantity(product.id, product.discountedPrice)}
+        >
           add to cart
         </button>
         <div>{renderReviews()}</div>
